[Refactor] Divide elements to components (e.g., SearchBar, Item, etc,.)
