+++ minjunseong/week_1-2/src/pages/BookCart.tsx
@@ -1,8 +1,8 @@
-// import { useState, useEffect } from "react";
 import styled from "styled-components";
 import { FaHome } from "react-icons/fa";
 import { useNavigate } from "react-router-dom";
 import { useEffect, useState } from "react";
+import BookCartItem from "../components/BookCartItem";
 
 interface BookType {
   title: string;
@@ -10,16 +10,6 @@ interface BookType {
   image: string;
   description: string;
 }
-
-// const BookImage = styled.img`
-//   width: 180px;
-//   height: 270px;
-// `;
-
-// 지금 해야하는 거 :
-// 홈 버튼 / 내 서재 label component grouping
-// 하단 viewer
-// 위 두개를 적절한 배치
 
 const BookCart = () => {
   const navigate = useNavigate();
@@ -30,12 +20,17 @@ const BookCart = () => {
 
   const [savedBooks, setSavedBooks] = useState<BookType[]>([]);
   useEffect(() => {
-    const data = JSON.parse(localStorage.getItem("myLibrary") || "[]");
-    setSavedBooks(data);
+    try {
+      const data = JSON.parse(localStorage.getItem("myLibrary") || "[]");
+      setSavedBooks(data);
+    } catch (err) {
+      console.error("잘못된 JSON 형식", err);
+      setSavedBooks([]);
+    }
   }, []);
 
   return (
-    <ViewContainer>
+    <div>
       <HeaderBar>
         <HomeButton onClick={handleToBookSelect}>
           <FaHome size="20" />
@@ -44,29 +39,13 @@ const BookCart = () => {
       </HeaderBar>
       <BookCartViewer>
         {savedBooks.map((book, index) => (
-          <BookItem key={index}>
-            <BookImage src={book.image} alt={book.title} />
-            <BookTitle>{book.title}</BookTitle>
-          </BookItem>
+          <BookCartItem key={index} book={book} />
         ))}
       </BookCartViewer>
-    </ViewContainer>
+    </div>
   );
 };
 
-// 근데 이 ViewContainer를 어차피 BookSelect에서도 사용하는데
-// 이걸 따로 빼서 import해서 쓰는게 좋을까?
-// 아니면 App.tsx에서 전역으로 설정해주는게 좋을까?
-const ViewContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100vw;
-  height: 100vh;
-`;
-
-// 헤더영역
 const HeaderBar = styled.div`
   width: 900px;
   display: flex;
@@ -76,7 +55,6 @@ const HeaderBar = styled.div`
   padding: 20px;
 `;
 
-// 좌측 홈버튼
 const HomeButton = styled.button`
   width: 50px;
   height: 50px;
@@ -92,7 +70,6 @@ const HomeButton = styled.button`
   }
 `;
 
-// 우측 라벨
 const TitleLabel = styled.div`
   width: 400px;
   height: 50px;
@@ -108,42 +85,15 @@ const TitleLabel = styled.div`
 const BookCartViewer = styled.div`
   width: 900px;
   display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  width: 870px;
+  flex-wrap: wrap;
+  justify-content: flex-start;
+  align-items: flex-start;
+  gap: 20px;
   height: 700px;
   border-radius: 30px;
   background-color: #d9d9d9;
-`;
-
-const BookItem = styled.div`
-  width: 180px;
-  height: 270px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  border-radius: 30px;
-  background-color: white;
-  shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
-  padding: 10px;
-  margin: 10px;
-`;
-
-const BookImage = styled.img`
-  max-width: 50px;
-  max-height: 50px;
-`;
-
-const BookTitle = styled.h3`
-  font-size: 16px;
-  font-weight: bold;
-  text-align: center;
-  max-width: 50px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
+  padding: 20px;
+  overflow-y: auto;
 `;
 
 export default BookCart;
